refactor: email step 과정 통합
